optimize: abstract getCaht function
