+++ server/middleware/auth.js
@@ -9,7 +9,7 @@ const auth = (req, res, next) => {
 
 const authAdmin = (req, res, next) => {
   if (req.user && req.user.isAdmin) return next()
-  return res.status(403).json({ status: 'error', message: 'permission denied' })
+  return res.status(403).json({ status: 'error', message: 'Permission denied' })
 }
 module.exports = {
   auth,
